keep pointer in window when held
If we shrink the window to fit its newly tiled place when dragging, it
looks ugly if the pointer ends up outside the window.  This ensures that
the held window is always placed so that it contains the pointer, and it
prevents the warp function from trying to override our pointer
placement.