place and resize window when managing
If we map and unmap for showhide(), then the window doesn't need to be
moved at all.  Move it into place when we initially manage it.  This
would be handled by arrange(), but we need it here for floating windows.

On the other hand, the XMapWindow will be covered by showhide, so we
don't need to map explicitly when managing a new window.  This may fix
focus-flicker when a window is created which, by rules, isn't visible to
start.